requirements.txt
transformers==4.30.0
datasets==2.12.0
torch==2.0.0
numpy==1.24.0
scikit-learn==1.2.0
faiss-cpu==1.7.3
tqdm==4.65.0
backports.lzma==0.0.14